rewrite kurun command to use kurun-server
And also overhaul the codebase because it was too oldschool
