--- example/test.go
@@ -1,34 +1,0 @@
-package main
-
-import (
-	"fmt"
-	"os"
-
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/client-go/kubernetes"
-	"k8s.io/client-go/rest"
-)
-
-func main() {
-	fmt.Println(os.Args)
-
-	config, err := rest.InClusterConfig()
-	if err != nil {
-		panic(err)
-	}
-
-	client, err := kubernetes.NewForConfig(config)
-	if err != nil {
-		panic(err)
-	}
-
-	nodes, err := client.CoreV1().Nodes().List(metav1.ListOptions{})
-	if err != nil {
-		panic(err)
-	}
-
-	fmt.Println("List of Kubernetes nodes:")
-	for _, node := range nodes.Items {
-		fmt.Printf("- %s - %s\n", node.Name, node.Labels)
-	}
-}
