+++ internal/cmd/common.go
@@ -1,0 +1,3 @@
+package cmd
+
+import _ "k8s.io/client-go/plugin/pkg/client/auth"
